fix: print format and delete unused cats task3.1 HW3

+++ HW3/task3/task3.c
@@ -18,7 +18,7 @@ int main() {
 
     read_input(&a, &b);
 
-    dif = (long)a - b;
+    dif = a - b;
 
-    printf("%d", dif);
+    printf("%ld", dif);
 }
